requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
boto3==1.14.47
botocore==1.17.47
django-rest-framework==0.1.0
django-storages==1.9.1
django==3.1
djangorestframework==3.11.1; python_version >= '3.5'
docutils==0.15.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
gunicorn==20.0.4
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
pillow==7.2.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2020.1
s3transfer==0.3.3
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transliterate==1.10.2
urllib3==1.25.10; python_version != '3.4'
whitenoise==5.2.0
